+++ credtools/wrappers/__init__.py
@@ -8,4 +8,5 @@ from credtools.wrappers.mesusie import run_mesusie
 from credtools.wrappers.multisusie import run_multisusie
 from credtools.wrappers.RSparsePro import run_rsparsepro
 from credtools.wrappers.susie import run_susie
+from credtools.wrappers.susie_ash import run_susie_ash
 from credtools.wrappers.susiex import run_susiex